定数とcanvas関係の関数をファイルに分ける
